Progress towards moving the chat card rendering to the rendering layer.
Refactoring done towards structuring data that gets put into a handlebars template, which does all the HTML for us.

I tried to get the existing system CSS to work but ultimately had to duplicate it; the existing card is so overly-classed and the CSS is overly specific and would just reproduce the crimes of the existing html.